+++ .idea/misc.xml
@@ -9,9 +9,11 @@
         <entry key="..\:/AppBuscaTutor/app/src/main/res/menu/my_menu.xml" value="0.1" />
         <entry key="..\:/Users/BRUNO/AndroidStudioProjects/demo_my_paypal/AppBuscaTutor/app/src/main/res/layout/activity_paypal.xml" value="0.1094224924012158" />
         <entry key="..\:/Users/BRUNO/AndroidStudioProjects/demo_my_paypal/AppBuscaTutor/app/src/main/res/layout/activity_verificacion_paypal.xml" value="0.21145833333333333" />
-        <entry key="..\:/Users/Miguel/AndroidStudioProjects/Ramas/gviernes_grupo1_backend_fisi/app/src/main/res/drawable/ic_search_tutor.xml" value="0.36574074074074076" />
-        <entry key="..\:/Users/Miguel/AndroidStudioProjects/Ramas/gviernes_grupo1_backend_fisi/app/src/main/res/layout/activity_buscar_tutor.xml" value="0.36594202898550726" />
-        <entry key="..\:/Users/Miguel/AndroidStudioProjects/Ramas/gviernes_grupo1_backend_fisi/app/src/main/res/layout/activity_item_tutor.xml" value="0.36594202898550726" />
+        <entry key="..\:/Users/BRUNO/AndroidStudioProjects/gviernes_grupo1_backend_fisi/app/src/main/res/layout/activity_homepage_tutores.xml" value="0.33" />
+        <entry key="..\:/Users/BRUNO/AndroidStudioProjects/gviernes_grupo1_backend_fisi/app/src/main/res/layout/activity_item_tutor.xml" value="0.26979166666666665" />
+        <entry key="..\:/Users/BRUNO/AndroidStudioProjects/gviernes_grupo1_backend_fisi/app/src/main/res/layout/recycler_template.xml" value="0.33" />
+        <entry key="..\:/Users/BRUNO/AndroidStudioProjects/projecto_movil/gviernes_grupo1_backend_fisi/app/src/main/res/layout/activity_registrar_datos_tutor.xml" value="0.20520833333333333" />
+        <entry key="..\:/Users/BRUNO/AndroidStudioProjects/projecto_movil/gviernes_grupo1_backend_fisi/app/src/main/res/layout/activity_verificacion_paypal.xml" value="0.25" />
         <entry key="..\:/Users/avira/AndroidStudioProjects/AppBuscaTutor/app/src/main/res/drawable/border.xml" value="0.21614583333333334" />
         <entry key="..\:/Users/avira/AndroidStudioProjects/AppBuscaTutor/app/src/main/res/drawable/border_foto.xml" value="0.21614583333333334" />
         <entry key="..\:/Users/avira/AndroidStudioProjects/AppBuscaTutor/app/src/main/res/drawable/bottom_nav_menu.xml" value="0.1" />
@@ -37,5 +39,12 @@
   </component>
   <component name="ProjectType">
     <option name="id" value="Android" />
+  </component>
+  <component name="VisualizationToolProject">
+    <option name="state">
+      <ProjectState>
+        <option name="scale" value="0.20520833333333333" />
+      </ProjectState>
+    </option>
   </component>
 </project>
