+++ app/src/main/res/layout/activity_verificacion_paypal.xml
@@ -48,6 +48,7 @@
                 android:layout_marginStart="30sp"
                 android:layout_marginLeft="30sp"
                 android:layout_marginTop="10sp"
+                android:layout_marginBottom="120sp"
                 android:layout_marginEnd="30sp"
                 android:layout_marginRight="30sp"
                 android:orientation="vertical">
@@ -69,17 +70,15 @@
                 <Button
                     android:id="@+id/id_btnA"
                     android:layout_width="match_parent"
-                    android:layout_height="10dp"
+                    android:layout_height="80sp"
                     android:layout_marginStart="80sp"
                     android:layout_marginEnd="80sp"
                     android:layout_marginTop="20sp"
-                    android:layout_marginBottom="190sp"
-                    android:layout_weight="1"
                     android:gravity="center"
                     android:text="@string/verificado_paypal_btn"
                     android:textStyle="bold"
                     app:backgroundTint="@color/Titulos"
-                    app:cornerRadius="20sp" />
+                    />
 
             </LinearLayout>
